requirements.txt
pymupdf>=1.23
pillow>=10.0
numpy>=1.24
